changing folder structure
