+++ client/src/app/(auth)/login.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import LogIn from "../../components/Auth/LoginScreen";
+import LogIn from "../../screens/auth/LoginScreen";
 
 const LoginPage = () => {
   return <LogIn />;
