+++ client/src/app/(home)/chatchannel.tsx
@@ -1,14 +1,14 @@
-import React from 'react'
-import { Stack } from 'expo-router';
-import ChatScreen from '../../components/Chat/ChatScreen'
-import SafeAreaWrapper from '../../components/Common/SafeAreaWrapper';
+import React from "react";
+import { Stack } from "expo-router";
+import ChatScreen from "../../screens/chat/ChatScreen";
+import SafeAreaWrapper from "../../components/common/SafeAreaWrapper";
 
 const ChatroomPage = () => {
-    return (
-      <SafeAreaWrapper>
-        <ChatScreen />
-      </SafeAreaWrapper>
-      );
-}
+  return (
+    <SafeAreaWrapper>
+      <ChatScreen />
+    </SafeAreaWrapper>
+  );
+};
 
-export default ChatroomPage
+export default ChatroomPage;
