+++ client/src/components/common/ChatMessage.tsx
