+++ client/src/components/common/ChatScreenFooter.tsx
