+++ client/src/components/common/MessageChannel.tsx
