+++ client/src/components/common/NearbyCount.tsx
