+++ client/src/components/common/SafeAreaWrapper.tsx
