+++ client/src/components/common/SignUpButton.tsx
