+++ client/src/contexts/SocketContext.tsx
@@ -2,7 +2,7 @@ import React, { createContext, useContext, useEffect, useState } from "react";
 import { io, Socket } from "socket.io-client";
 import { useLocation } from "./LocationContext";
 import { EXPO_IP } from "@env";
-import { AuthStore } from "../services/store";
+import { AuthStore } from "../services/AuthStore";
 
 const SocketContext = createContext<Socket | null>(null);
 
@@ -15,27 +15,25 @@ export const SocketProvider = ({ children }: { children: React.ReactNode }) => {
   const [mounted, setMounted] = useState(false);
   const locationContext = useLocation();
 
-
   useEffect(() => {
     const getToken = async () => {
       const token = await AuthStore.getRawState().userAuthInfo?.getIdToken();
       console.log("Token:", token);
       return token;
-    }
-    
+    };
+
     const initializeSocket = async () => {
       const token = await getToken();
       const socketIo = io(`http://${EXPO_IP}:8080`, {
         auth: {
           token: token,
-        }
+        },
       });
 
-      socketIo.connect()
+      socketIo.connect();
       setSocket(socketIo);
       setMounted(true);
-    }
-
+    };
 
     if (!mounted) {
       initializeSocket();
@@ -47,20 +45,18 @@ export const SocketProvider = ({ children }: { children: React.ReactNode }) => {
 
   // Listen to the socket state and run once the socket is set!
   useEffect(() => {
-
     if (!socket) return;
-    
+
     socket.on("connect", () => {
       console.log("Connected to server");
-      }
-    );
+    });
 
     return () => {
       console.log("[LOG]: Cleaning up sockets and mounted state.");
       socket.disconnect();
       setSocket(null);
       setMounted(false);
-    }
+    };
   }, [socket]);
 
   useEffect(() => {
